// src/app/interfaces/carriers.ts
export interface Carrier {
  id?:            number;
  Order_ID:       string;
  Vehicle?:        string;
  O_City?:         string;
  O_State?:       string;
  O_ZIP?:         string;
  D_City?:        string;
  D_State?:        string;
  D_ZIP?:          string;
  Carrier?:        string;
  Based?:          string;
  Phone1?:         string;
  Phone2?:         string;
  email?:          string;
  Ins_Expcode_ci?: string;
  ID?:             string;
  Last_Action?:    string;
  Deposit?:        string;
  Tariff?:         string;
  Carrier_Pay?:    string;
  Notes?:         string;

}